+++ NEWS.md
@@ -1,3 +1,11 @@
+# spant 2.6.0
+* Masked spectra plotted with image are now shown with a value of zero.
+* Added function to read .pta files - read_pulse_pta.
+* Added a function to make basis set objects from a directory of LCModel
+formatted RAW files - make_basis_from_raw.
+* Added horizontal line to fit result plots.
+* Improvements to reading LCM .raw files with variable numerical columns.
+
 # spant 2.5.0
 * Improved function documentation for water concentration scaling.
 * Added water and tCr scaling examples to the introduction vignette.